pyproject.toml
[tool.poetry]
name = "kanalservis"
version = "0.1.0"
description = ""
authors = ["Evgenii <jenek46@ya.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "~3.10"
django-cors-headers = "^3.14.0"
djangorestframework = "^3.14.0"
python-dotenv = "^1.0.0"
gunicorn = "^20.1.0"
aiohttp = "^3.8.4"
aiogram = "^2.25.1"
django = "^4.2rc1"
psycopg = {extras = ["binary"], version = "^3.1.8"}
markdown = "^3.4.3"
django-filter = "^22.1"
uvicorn = {extras = ["standard"], version = "^0.21.1"}


[tool.poetry.group.dev.dependencies]
python-dotenv = {extras = ["cli"], version = "^1.0.0"}
autopep8 = "^2.0.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
